# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Simple books library api",
    "description": "APIs for managing members, books and issuing of books to members",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Sandbox"
    }
  ],
  "paths": {
...
    "/issue": {
      "post": {
        "tags": [ "issue" ],
        "summary": "Issue a book",
        "description": "Specify the member id to whom the book is issued and the book id",
        "operationId": "create",
        "requestBody": {
          "description": "Details of issue book",
          "required": "true",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/addIssueRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/addIssueResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicErrorMessage"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicErrorMessage"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [ "issue" ],
        "summary": "Get details of book issues",
        "description": "Get details of book issues by specifying either a member id or a book id",
        "operationId": "getMultiple",
        "parameters": [ 
          {
            "in": "query",
            "name": "bookId",
            "type": "string",
            "description": "id of the book to be checked"
          },
          {
            "in": "query",
            "name": "memberId",
            "type": "string",
            "description": "id of the member to be checked"
          }
         ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/addIssueResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Record not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicErrorMessage"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicErrorMessage"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "addIssueRequest": {
        "type": "object",
        "required": [ "bookId", "memberId" ],
        "properties": {
          "bookId": {
            "type": "string",
            "description": "Id of the book to be issued" 
          },
          "memberId": {
            "type": "string",
            "description": "Id of the member to whom the book has to be issued" 
          }
        },
        "example": {
          "bookId": "book-id",
          "memberId": "member-id"
        }
      },
      "addIssueResponse": {
        "type": "object",
        "description": "Details of the member and the book",
        "properties": {
          "message": {
            "type": "string",
            "description": "Successful response"
          },
          "payload": {
            "type": "object",
            "description": "details of book and the member to whom it is issued",
            "properties": {
              "_id": {
                "type": "string",
                "description": "unique id representing issuinng of the book to the member"
              },
              "book": {
                "type": "object",
                "description": "Details of book",
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "unique id of the book"
                  },
                  "title": {
                    "type": "string",
                    "description": "book title"
                  },
                  "authors": {
                    "type": "array",
                    "description": "book authors",
                    "items": {
                      "type": "string",
                      "description": "author name"
                    }
                  }
                }
              },
              "member": {
                "type": "object",
                "description": "Details of member",
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "unique id of the member"
                  },
                  "firstName": {
                    "type": "string",
                    "description": "first name of the member"
                  },
                  "lastName": {
                    "type": "string",
                    "description": "last name of the member"
                  }
                }
              },
              "status": {
                "type": "string",
                "description": "book issue status"
              }
            }
          }
        },
        "example": {
          "_id": "issue-id",
          "book": {
            "id": "book-id",
            "title": "Jokes for dummies",
            "authors": [
              "W H Jenkins",
              "Clara C Turner"
            ]
          },
          "member": {
            "id": "member-id",
            "firstName": "Rajesh",
            "lastName": "Patil"
          },
          "status": "issued"
        }
      },
...
      "BasicErrorMessage": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        }
      }
    },
...
  }
}